Update gsnkit.adoc
Line 5, where it says, " you need to abstract gas for your users", how is "abstract" being used here? It is a noun, not a verb.

Line 47, where it says, "Also, the counter contract is going to naively pay for all the transactions that are submitted.", what do you mean by "naively"? This means like an innocent ignorance, like a child is naive about the world. Is this word used as jargon?